Player Interaction Update
Changed script to previous script I worked on a while back. I forgot it
is easier to just attach the object to the camera as a parent instead of
changing its position constantly. Your still going to want to mess with
it so that when you let go it does not spazz out. Also, the button to
pick up objects is now 'E'