Respawn Added
Added a quick and basic checkpoint respawning using a plane as a trigger
and a c# script. Checkpoint is based on a Vector3 coordinate, defined in
the update function and respawn after the player exits the trigger
plane. Still need to add the flash to white when entering the trigger
plane and making the trigger plane invisible