Lever animation added
Both levers in the cave can now be pulled. Also made it so that the key
and the three stones are able to be picked up now. Still having trouble
with triggering the cabin door with the key and I think I have it, but a
nullexceptionpointer is getting the better of me