Floating Tutorial Text, Highlights on MouseOver
In addition to these, I added some of the newer models and threw some
free textures on a few areas to give it a more complete look. We still
really need to texture everything ourselves and clean up a few more
models. Mitchell I added a drawbridge at the end of the third level,
just animate that with the last slot, so that is comes down and lets the
player enter the "Sanctuary"